Address comments; update tests
