+++ packages/node/test/integration/cross-node-multisig-creation.spec.ts
@@ -65,47 +65,38 @@ describe("Node can create multisig, other owners get notified", () => {
     firebaseServiceFactory.closeServiceConnections();
   });
 
-  it("Node A can create multisig and sync with Node B on new multisig creation", async () => {
-    const multisigAddress = await getNewMultisig(nodeA, [
-      nodeA.publicIdentifier,
-      nodeB.publicIdentifier
-    ]);
-    const openChannelsNodeA = await getChannelAddresses(nodeA);
-    const openChannelsNodeB = await getChannelAddresses(nodeB);
-    expect(openChannelsNodeA[0]).toEqual(multisigAddress);
-    expect(openChannelsNodeB[0]).toEqual(multisigAddress);
-  });
+  describe("Queued channel creation", () => {
+    it("Node A can create multiple back-to-back channels with Node B", async () => {
+      expect.assertions(5);
 
-  it("Node A can create multiple back-to-back channels with Node B", async () => {
-    expect.assertions(5);
+      let multisigAddress1: string;
+      let multisigAddress2: string;
+      const multisigAddress1Promise = getNewMultisig(nodeA, [
+        nodeA.publicIdentifier,
+        nodeB.publicIdentifier
+      ]);
+      const multisigAddress2Promise = getNewMultisig(nodeA, [
+        nodeA.publicIdentifier,
+        nodeB.publicIdentifier
+      ]);
 
-    let multisigAddress1: string;
-    let multisigAddress2: string;
-    const multisigAddress1Promise = getNewMultisig(nodeA, [
-      nodeA.publicIdentifier,
-      nodeB.publicIdentifier
-    ]);
-    const multisigAddress2Promise = getNewMultisig(nodeA, [
-      nodeA.publicIdentifier,
-      nodeB.publicIdentifier
-    ]);
+      // multisigAddress2Promise should have been blocked on
+      // multisigAddress1Promise finishing
+      multisigAddress1Promise.then(address => {
+        expect(multisigAddress2).toBeUndefined();
+      });
 
-    // multisigAddress2Promise should have been blocked on
-    // multisigAddress1Promise finishing
-    multisigAddress1Promise.then(address => {
-      expect(multisigAddress2).toBeUndefined();
+      multisigAddress2 = await multisigAddress2Promise;
+      // multisigAddress1Promise has already resolved by this point, simply
+      // getting its value
+      multisigAddress1 = await multisigAddress1Promise;
+
+      const openChannelsNodeA = await getChannelAddresses(nodeA);
+      const openChannelsNodeB = await getChannelAddresses(nodeB);
+      expect(openChannelsNodeA[0]).toEqual(multisigAddress1);
+      expect(openChannelsNodeB[0]).toEqual(multisigAddress1);
+      expect(openChannelsNodeA[1]).toEqual(multisigAddress2);
+      expect(openChannelsNodeB[1]).toEqual(multisigAddress2);
     });
-
-    multisigAddress2 = await multisigAddress2Promise;
-    // multisigAddress1Promise has already resolved by this point, simply
-    // getting its value
-    multisigAddress1 = await multisigAddress1Promise;
-
-    const openChannelsNodeA = await getChannelAddresses(nodeA);
-    const openChannelsNodeB = await getChannelAddresses(nodeB);
-    expect(openChannelsNodeA[0]).toEqual(multisigAddress1);
-    expect(openChannelsNodeB[0]).toEqual(multisigAddress1);
-    expect(openChannelsNodeA[1]).toEqual(multisigAddress2);
-    expect(openChannelsNodeB[1]).toEqual(multisigAddress2);
   });
 });
